refactor: dto 명칭 및 패키지 분류 명확화
