--- src/main/java/chess/dto/ErrorResponseDto.java
@@ -1,14 +1,0 @@
-package chess.dto;
-
-public class ErrorResponseDto {
-
-    private final String errorMessage;
-
-    public ErrorResponseDto(String errorMessage) {
-        this.errorMessage = errorMessage;
-    }
-
-    public String getErrorMessage() {
-        return errorMessage;
-    }
-}
