+++ src/main/java/chess/dto/response/ErrorResponseDto.java
@@ -1,0 +1,49 @@
+package chess.dto.response;
+
+import chess.domain.game.Score;
+import chess.domain.piece.Team;
+
+public class ErrorResponseDto {
+
+    private final String errorMessage;
+
+    public ErrorResponseDto(String errorMessage) {
+        this.errorMessage = errorMessage;
+    }
+
+    public String getErrorMessage() {
+        return errorMessage;
+    }
+
+    public static class StatusResponseDto {
+
+        private final double whiteScore;
+        private final double blackScore;
+        private final String winningTeam;
+
+        private StatusResponseDto(final double whiteScore, final double blackScore, final String winningTeam) {
+            this.whiteScore = whiteScore;
+            this.blackScore = blackScore;
+            this.winningTeam = winningTeam;
+        }
+
+        public static StatusResponseDto of(final Score score) {
+            double whiteScore = score.calculateScore(Team.WHITE);
+            double blackScore = score.calculateScore(Team.BLACK);
+            return new StatusResponseDto(whiteScore, blackScore,
+                score.calculateWinningTeam(whiteScore, blackScore).getValue());
+        }
+
+        public double getWhiteScore() {
+            return whiteScore;
+        }
+
+        public double getBlackScore() {
+            return blackScore;
+        }
+
+        public String getWinningTeam() {
+            return winningTeam;
+        }
+    }
+}
