+++ src/test/java/chess/service/ChessServiceTest.java
@@ -4,9 +4,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import chess.dto.StatusDto;
 import chess.dto.request.MoveRequestDto;
 import chess.dto.request.RoomRequestDto;
+import chess.dto.response.ErrorResponseDto.StatusResponseDto;
 import chess.dto.response.GameResponseDto;
 import chess.dto.response.RoomResponseDto;
 import chess.dto.response.RoomsResponseDto;
@@ -125,7 +125,7 @@ class ChessServiceTest {
     @Test
     void createStatus() {
         final Long id = createTestRoom("체스 초보만").getId();
-        final StatusDto status = chessService.createStatus(id);
+        final StatusResponseDto status = chessService.createStatus(id);
 
         assertThat(status.getBlackScore()).isEqualTo(38);
     }
